requirements.txt
aiohttp>=3.9.0
textual>=0.40.0
pyperclip>=1.8.2
lxml>=4.9.3
PyYAML>=6.0.1
